src/server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
Werkzeug==3.0.1
opencv-python-headless==4.9.0.80
pillow==10.2.0
ultralytics==8.1.0
azure-cognitiveservices-vision-computervision==0.9.0
google-generativeai==0.3.2
python-dotenv==1.0.1
matplotlib==3.8.2
numpy==1.26.4
